requirements.txt
future~=0.18.3
keyboard~=0.13.5
pyperclip~=1.8.2
playsound~=1.3.0
Pillow~=10.1.0
PyAutoGUI~=0.9.54
pynput~=1.7.6